requirements.txt
google-genai
pymupdf
python-dotenv
rich
json_repair